Move FootnotePreview to its own file
